#17 comment Classifier and move TileOverlay to a new file
